requirements.txt
# LLM providers (at least one required)
anthropic>=0.34.0
openai>=1.40.0         # optional — only needed if using OpenAI

# SSH for VPS communication (local machine only)
paramiko>=3.4.0

# HTTP for AnkiConnect (local machine only)
requests>=2.31.0

# Config & env
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Optional: web research enrichment
duckduckgo-search>=6.2.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
